Fix the initial download of archive points
Even when we specify an interval, if we set the start to now - archive
length, the data is returned from the larger -- lower resolution --
archive.

Work around this by setting the start to now - archive + interval, where
now = the end_time of a request for updates from the future (max_float).

We also ask the graph to flow `ToX to ensure that the desired window
remains on the screen.

Signed-off-by: David Scott <dave.scott@citrix.com>